move 'get organizers info' functions to core/utils.py
