+++ core/management/commands/new_event.py
@@ -7,11 +7,11 @@ from django.template.loader import render_to_string
 from slacker import Error as SlackerError
 
 from ...command_helpers import gather_event_date_from_prompt
-from ...forms import AddOrganizerForm, EventForm
+from ...forms import EventForm
 from ...models import Event
 from pictures.models import StockPicture
 from ...slack_client import slack
-from ...utils import get_coordinates_for_city
+from ...utils import get_coordinates_for_city, get_main_organizer, get_team, create_users
 
 DELIMITER = "\n-------------------------------------------------------------\n"
 
@@ -45,63 +45,6 @@ def get_basic_info():
         city, country, date))
 
     return (city, country, date, url, event_mail)
-
-
-def get_main_organizer():
-    """
-        We're asking user for name and address of main organizer, and return
-        a list of dictionary.
-    """
-    team = []
-    click.echo("Now let's talk about the team. First the main organizer:")
-    main_name = click.prompt(click.style(
-        "First and last name", bold=True, fg='yellow'))
-    main_email = click.prompt(click.style(
-        "E-mail address", bold=True, fg='yellow'))
-
-    team.append({'name': main_name, 'email': main_email})
-
-    click.echo(u"All right, the main organizer is {0} ({1})".format(
-        main_name, main_email))
-
-    return team
-
-
-def get_team(team):
-    """
-        We're asking user for names and address of the rest of the team,
-        and append that to a list we got from get_main_organizer
-    """
-    add_team = click.confirm(click.style(
-        "Do you want to add additional team members?", bold=True, fg='yellow'), default=False)
-    i = 1
-    while add_team:
-        i += 1
-        name = click.prompt(click.style(
-            "First and last name of #{0} member".format(i), bold=True, fg='yellow'))
-        email = click.prompt(click.style(
-            "E-mail address of #{0} member".format(i), bold=True, fg='yellow'))
-        if len(name) > 0:
-            team.append({'name': name, 'email': email})
-            click.echo("All right, the #{0} team member of Django Girls is {1} ({2})".format(
-                i, name, email))
-        add_team = click.confirm(click.style(
-            "Do you want to add additional team members?", bold=True, fg='yellow'), default=False)
-
-    return team
-
-
-def create_users(team, event):
-    """
-        Create or get User objects based on team list
-    """
-    members = []
-    for member in team:
-        member['event'] = event.pk
-        form = AddOrganizerForm(member)
-        user = form.save()
-        members.append(user)
-    return members
 
 
 def brag_on_slack_bang(city, country, team):
